% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaMME.R
\docType{package}
\name{AlphaMME}
\alias{AlphaMME}
\alias{AlphaMME-package}
\title{Alpha Mixed Model Equations}
\description{
This package is designed for genomic selection. It contains
several mixed model equation solvers from other R packages.
The solvers have been rewritten in C++ and streamlined for
increased speed and more efficient memory usage.
}
